requirements.txt
torch
tqdm
pillow
transformers
timm
einops